// src/app/shared/interfaces/Specie/specie.ts
export interface Specie {
  average_height: string;
  average_lifespan: string;
  classification: string;
  designation: string;
  eye_colors: string;
  filmIds: number[];
  films: string[];
  hair_colors: string;
  homeworldId: number;
  homeworld: string;
  id: number;
  language: string;
  name: string;
  peopleIds: number[];
  people: string[];
  skin_colors: string;
  url: string;
}